requirements.txt
ibm-watson>=4.2.1
certifi==2017.7.27.1
chardet==3.0.4
pandas
tqdm==4.42.1
click==6.7
clickclick==1.2.2
connexion==1.1.15
decorator==4.1.2
python-dotenv==0.11.0
Flask==0.12.2
Flask-Cors==3.0.3
gevent==1.2.2
greenlet==0.4.12
idna==2.6
inflection==0.3.1
injector==0.13.1
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
MarkupSafe==1.0
PyYAML==3.12
six==1.10.0
swagger-spec-validator==2.1.0
typing==3.6.2
Werkzeug==0.12.2
jsonschema==2.6.0
elasticsearch==7.5.1
